# Conflicts:
#   src/db/schemas/project.js

+++ src/routers/educationRouter.js
@@ -4,8 +4,7 @@ import { educationService } from '../services/educationService';
 import asyncHandler from '../utils/asyncHandler';
 
 const educationRouter = Router();
-// 기본적으로 유저 아이디 URI로 라우팅, 본인이든 다른 유저이든
-// 변수명이 userId 인 이유는 각 education마다 id속성이 있기때문에 구분하기 위해
+// 다른 유저의 정보 얻는 부분 필요. req에 userId값 필요
 educationRouter.get(
     '/education/:userId',
     login_required,
@@ -18,32 +17,26 @@ educationRouter.get(
         res.status(200).send(educationList);
     })
 );
-// 학력 추가는 post 메소드로 /add 로 보냄. education = {UserId,학교명,전공,졸업상태}
+// 학력 추가는 post 메소드, education = {UserId,학교명,전공,졸업상태}
 educationRouter.post(
-    '/education/:userId/add',
+    '/education',
     login_required,
     asyncHandler(async (req, res, next)=>{
-        const { userId } = req.params;
-        if(userId != req.currentUserId){
-            throw new Error('권한이 없습니다');
-        };
+        const userId = req.currentUserId;
         const { schoolName, major, graduationTypeCode } = req.body;
         const education = { userId, schoolName, major, graduationTypeCode };
         const addNewEducation = await educationService.addEducation({education});
         res.status(201).send(addNewEducation);
     })
 );
-// 변경은 put 메소드로 /edit 으로 education = { _id(학력),학교명,전공,졸업상태}
+// 변경은 patch 메소드, education = { _id(학력),학교명,전공,졸업상태}
 educationRouter.patch(
-    '/education/:userId/edit',
+    '/education',
     login_required,
     asyncHandler(async (req, res, next)=>{
-        const { userId } = req.params;
-        if(userId != req.currentUserId){
-            throw new Error('권한이 없습니다');
-        };
+        const userId = req.currentUserId;
         const { _id, schoolName, major, graduationTypeCode } = req.body;
-        const education = { _id, schoolName, major, graduationTypeCode };
+        const education = { userId, _id, schoolName, major, graduationTypeCode };
         const editEducation = await educationService.editEducation({education});
         if (editEducation.errorMessage){
             throw new Error(editEducation.errorMessage);
@@ -51,17 +44,14 @@ educationRouter.patch(
         res.status(200).send(editEducation);
     })
 );
-// 삭제는 delete 메소드로 /delete 로
+// 삭제는 delete 메소드
 educationRouter.delete(
-    '/education/:userId/delete',
+    '/education/:_id',
     login_required,
     asyncHandler(async (req, res, next)=>{
-        const { userId } = req.params;
-        if(userId != req.currentUserId){
-            throw new Error('권한이 없습니다');
-        };
-        const { _id } = req.body;
-        const deleteEducation = await educationService.removeEducation({_id});
+        const userId = req.currentUserId;
+        const { _id } = req.params;
+        const deleteEducation = await educationService.removeEducation({ userId, _id });
         if (deleteEducation.errorMessage){
             throw new Error(deleteEducation.errorMessage);
         };
